requirements.txt
flask
flask-cors
python-dotenv
openai
pymupdf
rembg
Pillow
gunicorn
